requirements.txt
Django>=5.2,<6.0
Pillow>=10.0.0
python-dotenv>=1.0.0